more work on attempting to get energy into a block

+++ src/main/kotlin/com/refinedmods/refinedstorage/api/network/INetwork.kt
@@ -26,7 +26,7 @@ interface INetwork {
     /**
      * @return the energy usage per tick of this network
      */
-    val energyUsage: Int
+    val energyUsage: Double
 
     /**
      * @return the energy storage
